Add mTLS (x509) authentication method
Note: I added a script: `scripts/setup_devel_env` which generate certs
and configuration if they don't exist. It may prove useful in trying
to test this change. This script configures against samltest.id, and
generates a unique SP entityId, so that turnpike developers don't
overwrite each other's configs accidentally.

The approach I took was to go ahead and have nginx perform the
authentication (verifying the client cert against a CA), and then pass
the client cert subject and issuer to the turnpike service for
authorization logic.

Also, note that I changed the auth order to try x509 first. The service
falls back to SAML if the x509 header is not present. In this way, an
endpoint can support either authentication method (SAML for humans,
mTLS for services).

`x509` auth takes a Python expression which evaluates to `True` or
`False`. It is passed a dictionary `x509` which contains two
attributes: `subject_dn` and `issuer_dn`, which can be used to further
restrict which certs can be used (nginx already verifies the trust
chain based on the configured CA file).

For example to restrict the endpoint to a certificate with the DN of
`CN=test`, you could use `x509['subject_dn'] == 'CN=test'`.

Note that CRL and/or OCSP support should be configured in
`NGINX_SSL_CONFIG` as needed.

To test, setup needed environment, start the services, and then execute

```
curl -k --cert services/nginx/certs/cert.pem --key services/nginx/certs/key.pem https://turnpike.example.com/api/turnpike/identity
```

and observe the resulting JSON document. Try a non-matching cert (any
old cert will do), and verify that it receives 400.

A cert which authenticates but fails the predicate will receive a 403.