Support x-rh-certauth-cn from upstream gateway
When the x-rh-certauth-cn and x-rh-certauth-issuer headers are passed,
pass them down.

For development purposes, I'm still configuring nginx to perform client
cert auth if it is passed.

Note that I noticed our upstream gateway's DN form starts with a '/', so
I prepended this when client cert auth is used.

I assume x-rh-certauth-cn is a misnomer, as it looks to be the whole DN,
starting with /CN=. It was likely named as CN since that is the only
field in a RHSM-registered system's DN.

If this assumption is wrong, or if we introduce an additional header for
the subject DN, should be a quick tweak to fix.

Note that with this change, the following variations all work
simultaneously:

- visit https://turnpike.example.com/api/turnpike/identity/ and login
  via SAML
- mimicking upstream TLS termination: `curl -k -H 'x-rh-certauth-cn: /CN=foo' -H 'x-rh-certauth-issuer: /CN=bar' https://turnpike.example.com/api/turnpike/identity/`
- having nginx do TLS termination: `curl -k --cert ./saml/certs/sp.crt --key ./saml/certs/sp.key https://turnpike.example.com/api/turnpike/identity/`